% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_indices.R
\name{add_indices}
\alias{add_indices}
\title{Add drought indices to a data.frame of climate variables}
\usage{
add_indices(x, scales = c(1, 2, 3, 6, 12), id = "id", lat = "lat",
  tmean = "tmean", tmin = "tmin", tmax = "tmax", prec = "prec",
  drop_lat = TRUE)
}
\arguments{
\item{x}{a data.frame with climate data, that has at least id for location
id, lat for latitude, year, month, tmean, tmin, and tmax for mean, minimum,
and maximum temperature, and prec for precipitation.}

\item{scales}{for which temporal scales should spei and spi be computed}

\item{id}{character string for location id variable}

\item{lat}{character string for lat variable}

\item{tmean}{character string for tmean variable}

\item{tmin}{character string for tmin variable}

\item{tmax}{character string for tmax variable}

\item{prec}{character string for prec variable}

\item{drop_lat}{should latitude be dropped from result?}
}
\value{
a data.frame consisting of the original climate data with additional
  variables for drought indices
}
\description{
Currently add DMI, and SPEI and SPI for different scales. PET for SPEI is
computed according to Hargreaves.
}
