% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{downscale}
\alias{downscale}
\title{Downscale monthly climate data to 30 arcseconds using Worldclim data}
\usage{
downscale(x, nc_path, wc_dir, varname, mode = "auto", parallel = TRUE)
}
\arguments{
\item{x}{a data.frame with at least id, lon, and lat}

\item{nc_path}{absolute path to the NetCDF file with monthly climate data}

\item{wc_dir}{absolute path to the WorldClim v2.0 TIF files for the
respective climate parameter}

\item{varname}{the name of the climate variable in the NetCDF file}

\item{mode}{either "auto", or one of "temp" or "prec"}

\item{parallel}{if TRUE (default), some parts of the code are run in parallel
for speed gain}
}
\value{
a tidy data.frame with id, year, month, and the extracted climate
  variable
}
\description{
Currently, CRU TS 4.01 is supported out of the box; for E-OBS > 18e, NetCDF
files with monthly resolution have to be produced first, e.g. using CDO
commands.
}
\references{
Mosier, T. M., Hill, D. F., & Sharp, K. V. (2014). 30-Arcsecond
  monthly climate surfaces with global land coverage. International Journal
  of Climatology, 34(7), 2175–2188. https://doi.org/10.1002/joc.3829
}
\keyword{manip}
