% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale_cru.R
\name{downscale_cru}
\alias{downscale_cru}
\title{Downscale CRU TS data to 30 arcseconds using Worldclim data}
\usage{
downscale_cru(x, cru_path, wc_dir, varname, mode = "auto",
  parallel = TRUE)
}
\arguments{
\item{x}{a data.frame with at least id, lon, and lat}

\item{cru_path}{absolute path to the NetCDF file with CRU climate data}

\item{wc_dir}{absolute path to the WorldClim v2.0 TIF files for the respective climate parameter}

\item{varname}{the name of the variable in the CRU TS NetCDF file}

\item{mode}{either "auto", or one of "temp" or "prec"}

\item{parallel}{if TRUE (default), some parts of the code are run in parallel for speed gain}
}
\value{
a tidy data.frame with id, year, month, and the extracted climate variable
}
\description{
Downscale CRU TS data to 30 arcseconds using Worldclim data
}
