% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate (or just extract) monthly climate data}
\usage{
interpolate(x, nc_path, varname, id = "id", ...)
}
\arguments{
\item{x}{a data.frame with at least id, lon, and lat}

\item{nc_path}{absolute path to the NetCDF file with monthly climate data}

\item{varname}{the name of the climate variable in the NetCDF file}

\item{id}{character string indicating name of column in x that holds unique
location ids}

\item{...}{additional arguments passed to \code{raster::extract()}}
}
\value{
a tidy data.frame with id, year, month, and the extracted climate
  variable
}
\description{
Interpolate (or just extract) monthly climate data
}
\keyword{manip}
